config: remove keylime.conf from the package
In site-packages/keylime we should not add any configuration file.  This
file is expected to be edited for, for example, to change the IP address
of the verifier.

For installations done by `pip`, the user needs to provide a
configuration file.  Currently this can be located in /usr/etc, /etc or
$HOME/.config.

Signed-off-by: Alberto Planas <aplanas@suse.com>